[FIX] web: pivot no content helper
This commit reverts [1], which fixed the no content
helper message of the pivot view, by moving the message
definition in the JS.

With Owl components, one must always use the translation
function in the env (env._t), and no longer import it from
'web.core'. Moreover, there is no valid reason to define the
message in JS here, except that there was a bug in owl, that
has been fixed by [2] (merged in Odoo by the previous commit).

[1] https://github.com/odoo/odoo/commit/6fa058c4ac9901cfe4afcf5ce24e3e1bcb86f6f3
[2] https://github.com/odoo/owl/commit/2af8cccd65424ea75fbaa9f0a83c6974be4b1b5d

closes odoo/odoo#51442

Signed-off-by: Géry Debongnie (ged) <ged@openerp.com>
Signed-off-by: Aaron Bohy (aab) <aab@odoo.com>