[FIX] account: allow manager to add bank account
The model account.setup.bank.manual.config (menu "Add a Brank
Account") was with the group account.group_account_user ("Show Full
Accounting Features").
Since 10d6b4255013, transient models have ACL and are not accessible
if the user does not belong in the group.
The account.setup.bank.manual.config model is triggered via the server
action account.action_new_bank_setting instead of a classical
menu. It's also displayed in an onboarding step of account.
Because of this, the group must be relaxed (and not just adding a
group of a view/action)

In saas-13.2, 10d6b4255013 made simply the wizard appears in readonly
but in saas-13.4, since 775dad97c7461e, trying to display a model on
which we do not have access produces and access rights error.

closes odoo/odoo#51330

X-original-commit: e3be515b954d7dc8401232bacf7b132a898b68d0
Signed-off-by: Christophe Monniez (moc) <moc@odoo.com>
Signed-off-by: Martin Trigaux (mat) <mat@odoo.com>