[FIX] hr: show Create employee button based on company employee
Before this commit, the `Create Employee` button was not visible if user is linked to employee in any other company as the domain was based on `employee_count`. (Which represents Employee belongs to all companies)

With this commit, we check for `employee_id` which makes sure if the user has linked Employee for currently activated company.

Followup on 9770d308e10bfccbb3dd61658cd0c7fcbc74d1f3

X-original-commit: cb9ac24caea53615e2808a36e850939a1ccaa041