[IMP] mail: display activity summary in dropdown
PURPOSE

Currently, the activity dropdown displays the activity types of activities.
The goal is to display the activity summaries instead, and fallback on the
activity type if there is no summary set.

SPECIFICATION

In the activity dropdown, display the activity summary instead of the activity
type. Only display the activity type if there is no summary set on the activity.
A long summary will be ended with a ellipsis.

Here, handled a long summary for chatter, list, kanban, and activity views.

LINKS

Task: 2254893
PR https://github.com/odoo/odoo/pull/51438

closes odoo/odoo#51438

Signed-off-by: Alexandre Kühn (aku) <aku@odoo.com>