[FIX] crm: apply threshold for leads/opportunities in merge wizard
Purpose of this task is to prevent merging of opportunities when
a regular (non super) user selects more than five records.

This commit limits total number of recorrds to be merged, up to maximum
five leads / opportunities for non superuser environment, and raises a
UserError otherwise explaining the threshold limit.

closes odoo/odoo#55024

Task-id: 2256126
X-original-commit: c25d7654f250c888cea84c20af16b2a32cc5b365
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>