[FIX] web: search panel: keep good active values
In saas-13.4, if a searchpanel_default_XXXX value was found in the
context to be used as the default value of a searchpanel category
(select="one" field), then the value shown to be active for that
category would always be that value even if another value would have
been selected at some point. This was due to the fact that the
activeValueId was reset at each reload of the category to be that
default value.

In saas-13.5, the problem has been solved by 631aa1e6918e061334f3b58705f62f6b67aab56c.
The present commit is simply the forward port of a test that covers the
situation described above.

closes odoo/odoo#56042

Task: 2259602
X-original-commit: 436765e4bfe8ad7e9dd6a812cfc0f5cae03cc351
Signed-off-by: Mathieu Duckerts-Antoine <Polymorphe57@users.noreply.github.com>