[FIX] account: apply correct filters
Before this commit, Clicking on "Unpaid Invoices" or  "Bills to Pay" will open records which includes Unposted moves as well while the count was only counting Posted + Unpaid moves only.

With this commit, We apply the default 'posted' filter, and include the partially paid invoices to make the filter behavior consistent with the count.

closes odoo/odoo#51763

X-original-commit: 81b7c017d87cdc5b03cfdc5ff7204b00f5f93599
Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>